requirements.txt
python-telegram-bot==21.3
aiohttp==3.9.5
aiofiles==23.2.1
python-dotenv==1.0.1
telethon==1.36.0
yt-dlp>=2024.1.0
